requirements.txt
APScheduler==3.7.0
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
idna==2.10
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.4
python-decouple>=3.4
dj-database-url>=0.5.0
django-cloudinary-storage==0.3.0
coverage>=5.5
